Did i just solved a 1600 rated problem?Mommy would be so proud of me

+++ 2074D.cpp
@@ -41,7 +41,7 @@ int main() {
                 mapy[x] = max(mapy[x], cnt);
             }
         }
-        int ans = 0 ;
+        long long ans = 0 ;
         for( auto x : mapy){
             ans += x.second ;
         }
